feat: Add support for filtering reactions by type
This commit modifies the `Reactable` trait to add a new parameter `$reactionType` to the `reactionsBy` method. This allows filtering the reactions by a specific type. The code also includes a check in the `Reacts` trait to delete the reaction if it already exists with the same type.

Note: This commit message follows the established convention of using a prefix to indicate the type of change (feat for feature, fix for bug fix, chore for maintenance tasks, etc.)